Wrote .md to .html converter
I can now write articles in an article.md file (same folder as the desired article.html) and run the .py script. Tested this on a few pages and it works well enough. The converter also adds the appropriate classes to the first h1 and paragraph. Some html elements cannot be easily expressed in markdown, but these can still be written as html within the .md file.